+++ lib/event_authentication.dart
@@ -3,5 +3,6 @@
 library event_authentication;
 
 export 'src/jwt.dart';
+export 'src/jwt_signing.dart';
 export 'src/login.dart';
 export 'src/user.dart';