+++ CHANGELOG.md
@@ -1,4 +1,11 @@
+## 0.1.1
+
+- Added noExpiry field that defaults to false to EmailLoginRequest
+- Added jwtSecret to AuthenticationSecretsRepository
+
 ## 0.1.0+1
+
+- Added login.dart to exports in event_authentication
 
 ## 0.1.0